+++ example/clj/todopipeline/steps.clj
@@ -14,12 +14,12 @@
 (def backend-repo "git@github.com:flosell/todo-backend-compojure.git")
 (def frontend-repo "git@github.com:flosell/todo-backend-client.git")
 
-(defn current-timestamp []
-  (time-format/unparse (time-format/formatters :date-hour-minute-second) (time/now)))
+(defn build-label [ctx]
+  (str (time-format/unparse (time-format/formatters :date) (time/now)) "." (:build-number ctx)))
 
 ; Let's try out some custom metadata:
 (defn set-build-name [args ctx]
-  (support/assoc-metadata! ctx :human-readable-build-name (current-timestamp))
+  (support/assoc-metadata! ctx :human-readable-build-label (build-label ctx))
   {:status :success})
 
 ;; This step does nothing more than to delegate to a library-function.