+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The official release will have a defined and more stable API. If you are already
 ## 0.13.0
 * Improvements: 
   * Added support for build-level metadata (#138). See [Build Metadata](https://github.com/flosell/lambdacd/wiki/Build%20Metadata) for details
+  * UI support for some kinds of metadata: 
+    * `:human-readable-build-label`
 * Bug fixes:
   * Catch Exception instead of Throwable in build steps to avoid catching Errors which cannot be handled (#148), thanks @hgsy!
 * Deprecated: